Add serialization format and IO
Saving profiling data by JLD or similar is fraught due to pointers,
MethodInstances, etc. This reconstructs a minimal StackFrame just by
saving the portable component. It's of course quite possible that line
numbers, etc, will not correspond to source code on some other machine
or at a later date. Even if we did save all the info needed to
reproduce the original environment, it's not clear there would be a
path to getting it on the user's system. So this form of portability
doesn't seem to be a high priority.

Note, however, that not all forms of colorization will be applicable
with serialized data.